[WIP] Initial import of libfdt for MultiROM.
This repository adds an Android-build system compatible
version of the libfdt library that is shipped with the
dtc tool. Source:
https://git.kernel.org/pub/scm/utils/dtc/dtc.git
This library enables the reading and manipulation of FDTs
appended to a kernel or boot image.

In the context of multirom, this library is used to
manipulate the FDT such that fstab entries can be removed
if needed. See the multirom repository for a more detailed
explanation.
TODO: link to mrom commit.

TO BE DISCUSSED:
* The library has been imported as a reduced version of
  the source repository since only the FDT processing and
  not the dts (source) is required. Moreover, the introduction
  of the include folder was required such that the android
  build system can export the libfdt header and make it
  accessible for other modules.
* Is a shared lib variant needed? (e.g., for host binaries)

Signed-off-by: Alexander Diewald <Diewi@diewald-net.com>